Integrate ip4s into Message

+++ core/src/main/scala/org/http4s/internal/parsing/Rfc3986.scala
@@ -80,7 +80,7 @@ private[http4s] object Rfc3986 {
   }
 
   val ipv4Address: Parser[Ipv4Address] =
-    ipv4Bytes.map { case (a, b, c, d) => Ipv4Address.fromBytes(a, b, c, d) }
+    ipv4Bytes.map { case (a, b, c, d) => Ipv4Address.fromBytes(a.toInt, b.toInt, c.toInt, d.toInt) }
 
   val ipv6Address: Parser[Ipv6Address] = {
     import cats.parse.{Parser => P}
